upload and separate test, delete old test file
